Rename `mtwt` to `hygiene`
