+++ src/librustc/session/config.rs
@@ -735,8 +735,8 @@ options! {DebuggingOptions, DebuggingSetter, basic_debugging_options,
           "for every macro invocation, print its name and arguments"),
     enable_nonzeroing_move_hints: bool = (false, parse_bool,
           "force nonzeroing move optimization on"),
-    keep_mtwt_tables: bool = (false, parse_bool,
-          "don't clear the resolution tables after analysis"),
+    keep_hygiene_data: bool = (false, parse_bool,
+          "don't clear the hygiene data after analysis"),
     keep_ast: bool = (false, parse_bool,
           "keep the AST after lowering it to HIR"),
     show_span: Option<String> = (None, parse_opt_string,
