+++ src/librustc_driver/driver.rs
@@ -236,8 +236,8 @@ pub fn compile_input(sess: &Session,
     Ok(())
 }
 
-fn keep_mtwt_tables(sess: &Session) -> bool {
-    sess.opts.debugging_opts.keep_mtwt_tables
+fn keep_hygiene_data(sess: &Session) -> bool {
+    sess.opts.debugging_opts.keep_hygiene_data
 }
 
 fn keep_ast(sess: &Session) -> bool {
@@ -479,7 +479,7 @@ pub fn phase_1_parse_input<'a>(sess: &'a Session,
                                input: &Input)
                                -> PResult<'a, ast::Crate> {
     // These may be left in an incoherent state after a previous compile.
-    syntax::ext::mtwt::reset_hygiene_data();
+    syntax::ext::hygiene::reset_hygiene_data();
     // `clear_ident_interner` can be used to free memory, but it does not restore the initial state.
     token::reset_ident_interner();
     let continue_after_error = sess.opts.continue_parse_after_error;
@@ -760,9 +760,9 @@ pub fn phase_2_configure_and_expand<'a, F>(sess: &Session,
         hir_map::Forest::new(lower_crate(sess, &krate, &mut resolver), &sess.dep_graph)
     });
 
-    // Discard MTWT tables that aren't required past lowering to HIR.
-    if !keep_mtwt_tables(sess) {
-        syntax::ext::mtwt::reset_hygiene_data();
+    // Discard hygiene data, which isn't required past lowering to HIR.
+    if !keep_hygiene_data(sess) {
+        syntax::ext::hygiene::reset_hygiene_data();
     }
 
     Ok(ExpansionResult {
